fix: merging digests with actions/upload-artifact@v4
v4 doesn't allow uploading to the same entry multiple times, so you need
to upload to separate ones and then merge them manually. In this case we
just glob and download all artifacts matching the pattern (as
recommended in the documentation).

To be able to use the matrix value in the artifact name we needed to
remove the slash.